test: 신규 단위 검증을 통합 회귀로 승격한다
